requirements.txt
librosa
tiktoken
pyaudio
yt_dlp
openai
tinygrad
pydub
google-api-python-client
pandas
python-dotenv
qdrant-client>=1.1.1
sentence-transformers